move all dsl classes into single file
